# docs/swagger.yaml
basePath: /api/v2
definitions:
  apperrors.ErrorResponse:
    properties:
      error_response:
        type: string
    type: object
...
  doc_structs.UserResponse:
    properties:
      user:
        $ref: '#/definitions/dto.UserPublicInfo'
    type: object
...
  dto.AuthInfo:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  dto.UserPublicInfo:
    properties:
      avatar_url:
        type: string
      description:
        type: string
      email:
        type: string
      name:
        type: string
      surname:
        type: string
      user_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: kdanil01@mail.ru
    name: Капитанов Даниил
    url: https://vk.com/poophead27
  description: Лучшее и единственно приложение, имитирующее Trello.
  license:
    name: None
    url: None
  title: LA TABULA API
  version: "2.0"
paths:
  /auth/login/:
    post:
      consumes:
      - application/json
      description: Для этого использует сессии
      parameters:
      - description: Эл. почта и логин пользователя
        in: body
        name: authData
        required: true
        schema:
          $ref: '#/definitions/dto.AuthInfo'
      produces:
      - application/json
      responses:
        "200":
          description: Объект пользователя
          schema:
            $ref: '#/definitions/doc_structs.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Войти в систему
      tags:
      - auth
  /auth/logout/:
    delete:
      consumes:
      - application/json
      description: Удаляет текущую сессию пользователя. Может сделать только авторизированный
        пользователь. Текущая сессия определяется по cookie "tabula_user", в которой
        лежит строка-токен.
      produces:
      - application/json
      responses:
        "204":
          description: no content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Выйти из системы
      tags:
      - auth
  /auth/signup/:
    post:
      consumes:
      - application/json
      description: Также вводит пользователя в систему
      parameters:
      - description: Базовые данные пользователя
        in: body
        name: signup
        required: true
        schema:
          $ref: '#/definitions/dto.AuthInfo'
      produces:
      - application/json
      responses:
        "200":
          description: Объект пользователя
          schema:
            $ref: '#/definitions/doc_structs.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Зарегистрировать нового пользователя
      tags:
      - auth
  /auth/verify:
    get:
      consumes:
      - application/json
      description: Узнать существует ли сессия текущего пользователя. Сессия определяется
        по cookie "tabula_user", в которой лежит строка-токен.
      produces:
      - application/json
      responses:
        "200":
          description: Объект пользователя
          schema:
            $ref: '#/definitions/doc_structs.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Подтвердить вход
      tags:
      - auth
...
swagger: "2.0"